requirements.txt
aiogram==3.10.0
openai>=1.50.0
faiss-cpu>=1.8.0
numpy>=1.26.4
python-docx>=1.1.2
pydantic-settings>=2.3.0
